Found a way to treat non-modded players as blocked.

+++ ds3runtime/scripts/ds3_iframe_patch.h
@@ -12,18 +12,18 @@ namespace ds3runtime::ds3IFramePatch {
 
 struct HitBoxRecord
 {
-	int64_t uniqueId;
-	int64_t parentUniqueId;
-	int64_t siblingUniqueId;
-	int64_t childUniqueId;
+	uint16_t uniqueId;
+	uint16_t parentUniqueId;
+	uint16_t siblingUniqueId;
+	uint16_t childUniqueId;
 };
 
 struct ReceivedHitRecord
 {
-	int64_t uniqueId;
-	int64_t parentUniqueId;
-	int64_t siblingUniqueId;
-	int64_t childUniqueId;
+	uint16_t uniqueId;
+	uint16_t parentUniqueId;
+	uint16_t siblingUniqueId;
+	uint16_t childUniqueId;
 	uint64_t timeStamp;
 };
 
@@ -45,24 +45,24 @@ public:
 
 	void unregisterOutgoingHitBox(uintptr_t hitBoxAddress);
 
-	void registerOutgoingHitBoxParent(uintptr_t hitBoxAddress, int64_t parentUniqueId);
+	void registerOutgoingHitBoxParent(uintptr_t hitBoxAddress, uint16_t parentUniqueId);
 
-	void registerOutgoingHitBoxChild(uintptr_t hitBoxAddress, int64_t childUniqueId);
+	void registerOutgoingHitBoxChild(uintptr_t hitBoxAddress, uint16_t childUniqueId);
 
-	void registerOutgoingHitBoxSibling(uintptr_t hitBoxAddress, int64_t siblingUniqueId);
+	void registerOutgoingHitBoxSibling(uintptr_t hitBoxAddress, uint16_t siblingUniqueId);
 
 	bool hasOutgoingHitBox(uintptr_t hitBoxAddress);
 
 	HitBoxRecord getOutgoingHitBoxRecord(uintptr_t hitBoxAddress);
 private:
-	int64_t outgoingUniqueIdCounter = 0;
+	uint16_t outgoingUniqueIdCounter = 0;
 	std::unordered_map<uintptr_t, HitBoxRecord> outgoingHitBoxIdMap;
 	std::unordered_map<uintptr_t*, std::vector<ReceivedHitRecord>> incomingHitBoxIdMap;
-	std::vector<int64_t> hitsToIgnoreThisFrame;
+	std::vector<uint16_t> hitsToIgnoreThisFrame;
 
 	bool doesReceivedHitShareUniqueId(ReceivedHitRecord hitRecord, ReceivedHitRecord otherHitRecord);
 
-	bool doesReceivedHitContainUniqueId(ReceivedHitRecord hitRecord, int64_t searchUniqueId);
+	bool doesReceivedHitContainUniqueId(ReceivedHitRecord hitRecord, uint16_t searchUniqueId);
 };
 
 }
